More comments added, minor changes
Renamed some functions, such as addTexture to setTexture. Added comments explaining TileProvider more. Fixed saving/loading until loading placed tileObjects is fully implemented.
The bug caused was Tile.TileType was given an invalid tileType ID, this was caused because of a misaligned pointer during loading, due to a not fully implemented feature still in progress.

+++ survivalGame/src/survivalGame/Direction.java
@@ -18,7 +18,7 @@ public enum Direction {
 		return mask;
 	}
 	
-	public byte getRotationID() {
+	public int getRotationID() {
 		return ID;
 	}
